Refactor: move inputs function to an object
